refactor: extract `_send_windows_request` to the `named_pipe_helper.py`
